backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langgraph
chromadb
python-dotenv
httpx
pytest
